refactor(client): improve mocking
