+++ client/src/tests/e2e/mocks/MockHelper.ts
@@ -4,18 +4,18 @@ import { HttpMethod } from "@core/services/khaos/KhaosTypes";
 export interface MockArgs<Fixture> {
   status: number;
   endpoint: string;
+  http: HttpMethod;
   fixture: Fixture;
-  httpMethod: HttpMethod;
 };
 
 export class MockHelper {
-  constructor(private page: Page) { };
+  constructor(private page: Page) {};
 
-  public async mock<F>({ status, endpoint, fixture, httpMethod }: MockArgs<F>): Promise<void> {
+  public async mock<F>({ endpoint, http, status, fixture }: MockArgs<F>) {
     await this.page.route(endpoint, async (route) => {
-      if (route.request().method() !== httpMethod) return;
+      if (route.request().method() !== http) return;
 
-      route.fulfill({
+      await route.fulfill({
         status: status,
         contentType: "application/json",
         body: JSON.stringify(fixture),
